login redirect added

+++ project.py
@@ -54,6 +54,9 @@ def restaurantMenu(restaurant_id):
 
 @app.route('/restaurant/<int:restaurant_id>/new/',methods=['GET','POST'])
 def newMenuItem(restaurant_id):
+	if 'username' not in login_session:
+		return redirect('/')
+
 	if request.method=='POST':
 		newItem=MenuItem(name=request.form['name'],restaurant_id=restaurant_id)
 		session.add(newItem)
@@ -67,6 +70,9 @@ def newMenuItem(restaurant_id):
 
 @app.route('/restaurant/<int:restaurant_id>/<int:menu_id>/edit/',methods=['GET','POST'])
 def editMenuItem(restaurant_id,menu_id):
+	if 'username' not in login_session:
+		return redirect('/')
+		
 	editedItem=session.query(MenuItem).filter_by(id=menu_id).one()
 	if request.method=='POST':
 		if request.form['name']:
@@ -81,6 +87,9 @@ def editMenuItem(restaurant_id,menu_id):
 
 @app.route('/restaurant/<int:restaurant_id>/<int:menu_id>/delete/',methods=['GET','POST'])
 def deleteMenuItem(restaurant_id,menu_id):
+	if 'username' not in login_session:
+		return redirect('/')
+		
 	deleteItem=session.query(MenuItem).filter_by(id=menu_id).one()
 	if request.method=='POST':
 		session.delete(deleteItem)
